реализация команд сервера
