--- src/main/java/Server/commands/Executable.java
@@ -1,5 +1,0 @@
-package Server.commands;
-
-public interface Executable {
-    boolean execution(String[] args);
-}
