+++ src/main/java/client/commands/Add.java
@@ -3,7 +3,7 @@ package client.commands;
 import common.Request;
 import common.obj.HumanBeing;
 import client.Ask;
-import client.Console;
+import client.console.Console;
 
 public class Add extends Command {
     private final Console console;
