+++ src/main/java/client/commands/FilterLessThanWeaponType.java
@@ -1,9 +1,8 @@
 package client.commands;
 
 import common.Request;
-import common.obj.HumanBeing;
 import common.obj.WeaponType;
-import client.Console;
+import client.console.Console;
 
 public class FilterLessThanWeaponType extends Command {
     private final Console console;
