+++ src/main/java/client/commands/PrintFieldDescendingWeaponType.java
@@ -1,8 +1,7 @@
 package client.commands;
 
 import common.Request;
-import common.obj.HumanBeing;
-import client.Console;
+import client.console.Console;
 
 public class PrintFieldDescendingWeaponType extends Command {
     private final Console console;
