+++ src/main/java/client/commands/UpdateID.java
@@ -1,7 +1,6 @@
 package client.commands;
 
-import client.Ask;
-import client.Console;
+import client.console.Console;
 import common.Request;
 
 public class UpdateID extends Command {
