+++ src/main/java/common/exeptions/NotBooleanExeption.java
@@ -1,3 +1,3 @@
-package exeptions;
+package common.exeptions;
 
 public class NotBooleanExeption extends Exception{}
