+++ src/main/java/server/commands/Executable.java
@@ -1,0 +1,7 @@
+package server.commands;
+
+import common.Request;
+
+public interface Executable {
+    boolean execution(Request request);
+}
