+++ src/main/java/server/commands/UpdateID.java
@@ -1,9 +1,9 @@
-package Server.commands;
+package server.commands;
 
 import client.commands.Command;
-import Server.managers.CollectionManager;
+import server.managers.CollectionManager;
 import client.Ask;
-import client.Console;
+import client.console.Console;
 
 public class UpdateID extends Command {
     private final Console console;
